CF-2 Renaming
