--- src/main/java/com/bsjhx/cashflow/application/bucket/command/BucketCommands.java
@@ -1,11 +1,0 @@
-package com.bsjhx.cashflow.application.bucket.command;
-
-import java.util.UUID;
-
-public interface BucketCommands {
-    record OpenBucketCommand(UUID bucketId) {
-    }
-
-    record TransferMoneyCommand(UUID bucketId, Double amount) {
-    }
-}
