+++ src/main/java/com/bsjhx/cashflow/application/tracksheet/command/MoneyTransferCommandHandler.java
@@ -1,7 +1,7 @@
-package com.bsjhx.cashflow.application.bucket.command;
+package com.bsjhx.cashflow.application.tracksheet.command;
 
 import com.bsjhx.cashflow.adapters.outbound.EventStore;
-import com.bsjhx.cashflow.application.bucket.command.BucketCommands.TransferMoneyCommand;
+import com.bsjhx.cashflow.application.tracksheet.command.TrackSheetCommands.TransferMoneyCommand;
 import com.bsjhx.cashflow.domain.tracksheet.TrackSheet;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,7 +13,7 @@ public class MoneyTransferCommandHandler {
     private final EventStore eventStore;
 
     public void handle(TransferMoneyCommand transferMoneyCommand) {
-        var pastEvents = eventStore.loadEvents(transferMoneyCommand.bucketId())
+        var pastEvents = eventStore.loadEvents(transferMoneyCommand.trackSheetId())
                 .orElseThrow(() -> new IllegalArgumentException("Bucket not found"));
 
         var bucket = TrackSheet.fromEvents(pastEvents);
