+++ src/main/java/com/bsjhx/cashflow/application/tracksheet/command/OpenBucketCommandHandler.java
@@ -1,7 +1,7 @@
-package com.bsjhx.cashflow.application.bucket.command;
+package com.bsjhx.cashflow.application.tracksheet.command;
 
 import com.bsjhx.cashflow.adapters.outbound.EventStore;
-import com.bsjhx.cashflow.application.bucket.command.BucketCommands.OpenBucketCommand;
+import com.bsjhx.cashflow.application.tracksheet.command.TrackSheetCommands.OpenTrackSheetCommand;
 import com.bsjhx.cashflow.domain.tracksheet.TrackSheet;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -14,9 +14,9 @@ public class OpenBucketCommandHandler {
 
     private final EventStore eventStore;
 
-    public void handle(OpenBucketCommand openBucketCommand) {
-        var pastEvents = eventStore.loadEvents(openBucketCommand.bucketId()).orElse(List.of());
-        var newBucket = TrackSheet.createNew(openBucketCommand.bucketId(), pastEvents);
+    public void handle(OpenTrackSheetCommand openBucketCommand) {
+        var pastEvents = eventStore.loadEvents(openBucketCommand.trackSheetId()).orElse(List.of());
+        var newBucket = TrackSheet.createNew(openBucketCommand.trackSheetId(), pastEvents);
         
         eventStore.saveEvents(newBucket.getId(), newBucket.getUncommittedEvents());
         newBucket.markChangesAsCommitted();
