+++ src/main/java/com/bsjhx/cashflow/application/tracksheet/command/TrackSheetCommands.java
@@ -1,0 +1,11 @@
+package com.bsjhx.cashflow.application.tracksheet.command;
+
+import java.util.UUID;
+
+public interface TrackSheetCommands {
+    record OpenTrackSheetCommand(UUID trackSheetId) {
+    }
+
+    record TransferMoneyCommand(UUID trackSheetId, Double amount) {
+    }
+}
